backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pycups==2.0.4
python-sane==2.9.1
zeroconf==0.132.2
websockets==12.0
python-multipart==0.0.9
aiofiles==23.2.1
Pillow==10.3.0
httpx==0.27.0
